+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
-# [0.0.1]
+## [0.0.1] - 2025-04-14
 
-- Created this extension!
+### Added
+
+- Integrated React-based custom editor for `.bru` files.
+- Added command `vs-bruno.openPlainText` to re-open `.bru` in text mode (F7).